Fix issue causing Nake temp files appear in a root directory
Nake uses Base64 encoding to generate temporary directory names. And
then combines them with a user defined TempDirectory via Path.Combine.
Since Base64 encoding allows '/' character, result of combining folders
might not be expected. For example:
  Path.Combine("C:\Temp", "/base64string");

will results in a '/base64string' folder omitting the root Temp folder.

Fixed it by replacing '/' character in a hash with '_' character. As it
occurs, it is used in some alternative implementations of Base64:
http://en.wikipedia.org/wiki/Base64#Variants_summary_table